Update civilization_first_build.py
Changed some things around, most important change is the population system. Once you upgrade your advertising level a random number is generated every 100 or so seconds and this number represents the amount of people that are looking to live in your town. The population_increment() function generates a random amount of people that want to purchase a house and it looks through the list of all the houses the player has bought to see if the town can accomodate them. That function is still a massive WIP and is nowehere near done, mostly in the testing stage.