Add files via upload
Did some graphical changes, fixed some bugs. Still working on implement all the necessary features. Check out the TO-DO list at the top of the file for more information on what is being worked on. The main GUI was changed a bit. But that may be changed a few more times, not sure yet.